feat: setup hitemp credentials each time its called
